Refactor workout code from parser, split filters
