+++ src/main/java/utility/Parser.java
@@ -3,12 +3,18 @@ package utility;
 import constants.ErrorConstant;
 import constants.HealthConstant;
 import constants.UiConstant;
+import constants.WorkoutConstant;
+import ui.Handler;
 import ui.Output;
 
 import health.Appointment;
 import health.Bmi;
 import health.HealthList;
 import health.Period;
+import workouts.Gym;
+import workouts.GymStation;
+import workouts.Run;
+import workouts.Workout;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -255,5 +261,105 @@ public class Parser {
             endIndex = input.length();
         }
         return input.substring(startIndex, endIndex).trim();
+    }
+
+    //@@author JustinSoh
+    /**
+     * Retrieves the date from the input for a Gym output.
+     * Returns empty string if not specified.
+     *
+     * @param input The user input string.
+     * @return A string representing the date.
+     */
+    public static String getDateFromGym(String input) {
+        try {
+            return extractSubstringFromSpecificIndex(input, WorkoutConstant.SPLIT_BY_DATE);
+        } catch (Exception e) {
+            return "";
+        }
+    }
+
+    /**
+     * Retrieves the number of gym stations in one Gym object from user input.
+     *
+     * @param input The user input string.
+     * @throws CustomExceptions.InsufficientInput If the user input is insufficient.
+     * @throws CustomExceptions.InvalidInput      If the user input is invalid or blank.
+     */
+
+    public static int getNumberOfGymStations(String input) throws CustomExceptions.InsufficientInput,
+            CustomExceptions.InvalidInput {
+        String numberOfStationString = extractSubstringFromSpecificIndex(input,
+                WorkoutConstant.SPLIT_BY_NUMBER_OF_STATIONS);
+        assert Integer.parseInt(numberOfStationString) > 0 : ErrorConstant.NEGATIVE_VALUE_ERROR;
+        return Integer.parseInt(numberOfStationString);
+    }
+
+    /**
+     * Retrieves the gym station details and adds a GymStation object to Gym.
+     *
+     * @param numberOfStations The number of stations in one gym session.
+     * @param gym              The Gym object.
+     */
+    public static void getGymStation(int numberOfStations, Gym gym) {
+        int i = 0;
+        while (i < numberOfStations) {
+            try {
+
+                Output.printGymStationPrompt(i + 1);
+                String userInput = Handler.in.nextLine();
+                GymStation.addGymStationInputValid(gym, userInput);
+                i++;
+            } catch (CustomExceptions.InsufficientInput | CustomExceptions.InvalidInput e) {
+                Output.printException(e.getMessage());
+            }
+        }
+        Output.printAddGym(gym);
+    }
+
+    /**
+     * Checks the type of exercise based on the user input.
+     * Usage: to use this method whenever the user enters a new exercise.
+     * Handles all the checks for input validity and sufficiency.
+     * Can assume input is valid and sufficient if no exceptions are thrown.
+     *
+     * @param userInput The user input string.
+     * @return The type of exercise {@code Constant.RUN} or {@code Constant.GYM}.
+     * @throws CustomExceptions.InvalidInput      If the user input is invalid or blank.
+     * @throws CustomExceptions.InsufficientInput If the user input is insufficient.
+     */
+    public static String checkTypeOfExercise(String userInput) throws
+            CustomExceptions.InvalidInput,
+            CustomExceptions.InsufficientInput {
+
+        boolean exerciseTypeIsValid = false;
+        boolean isRunValid = false;
+        boolean isGymValid = false;
+
+        String exerciseType = extractSubstringFromSpecificIndex(userInput,
+                WorkoutConstant.SPLIT_BY_EXERCISE_TYPE);
+
+        exerciseTypeIsValid = Workout.checkIfExerciseTypeIsValid(exerciseType);
+        boolean isRun = exerciseType.equals(WorkoutConstant.RUN);
+        boolean isGym = exerciseType.equals(WorkoutConstant.GYM);
+
+        if (isRun) {
+            String runDistance = extractSubstringFromSpecificIndex(userInput, WorkoutConstant.SPLIT_BY_DISTANCE);
+            String runTime = extractSubstringFromSpecificIndex(userInput, WorkoutConstant.SPLIT_BY_TIME);
+            String runDate = extractSubstringFromSpecificIndex(userInput, WorkoutConstant.SPLIT_BY_DATE);
+            isRunValid = Run.checkIfRunIsValid(runDistance, runTime, runDate);
+        } else if (isGym) {
+            String numberOfStations = extractSubstringFromSpecificIndex(userInput,
+                    WorkoutConstant.SPLIT_BY_NUMBER_OF_STATIONS);
+            isGymValid = Gym.checkIfGymIsValid(numberOfStations);
+        }
+
+        if (exerciseTypeIsValid && isRunValid) {
+            return WorkoutConstant.RUN;
+        } else if (exerciseTypeIsValid && isGymValid) {
+            return WorkoutConstant.GYM;
+        } else {
+            return "";
+        }
     }
 }
