+++ src/main/java/utility/WorkoutFilters.java
@@ -1,0 +1,4 @@
+package utility;
+
+public class WorkoutFilters {
+}
